Now using more specific errors in this class

+++ private/lib/Service/AuthenticationService.php
@@ -3,12 +3,14 @@
 namespace App\Service;
 
 use App\Database\Exception\NotFoundException;
-use App\Database\Model\User;
-use App\Database\Repository\UserRepository;
-use App\Exception\InvalidParameterException;
-use App\Utilities\Session;
+use App\Database\Models\User;
+use App\Database\Repositories\UserRepository;
+use App\Utility\Session;
+use InvalidArgumentException;
+use LengthException;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Cache\CacheItem;
+use Symfony\Contracts\Cache\CacheInterface;
 
 class AuthenticationService
 {
@@ -16,10 +18,11 @@ class AuthenticationService
     protected const SESSION_DATA_USER_ID = 'sessionUserId';
     protected const SESSION_DATA_USER_NAME = 'sessionUsername';
     protected const SESSION_DATA_PRIVILEGE_LEVEL = 'sessionPrivilegeLevel';
+    protected const DEFAULT_SESSION_EXPIRY_TIME = 86400; // 24 hours
 
     protected UserRepository $userRepository;
 
-    protected FilesystemAdapter $cache;
+    protected CacheInterface $cache;
 
     public function __construct()
     {
@@ -27,13 +30,13 @@ class AuthenticationService
         $this->cache          = new FilesystemAdapter('', 0, SESSION_CACHE_PATH);
     }
 
-    public function register(string $username, string $password, string $email): void
+    public function registerNewUser(string $username, string $password, string $email): void
     {
         $this->validateUsernameAndPasswordFields($username, $password);
 
         // Ensure that the provided email address is valid
         if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
-            throw new \RuntimeException('Provided email address is invalid');
+            throw new InvalidArgumentException('Provided email address is invalid');
         }
 
         $encryptedPassword = password_hash($password, PASSWORD_ARGON2ID);
@@ -41,13 +44,14 @@ class AuthenticationService
         $user = new User();
         $user->setUsername($username)
             ->setPassword($encryptedPassword)
-            ->setEmailAddress($email);
+            ->setEmailAddress($email)
+            ->setPrivilegeLevel(User::PRIVILEGE_LEVEL_USER);
 
         $this->userRepository->queue($user);
         $this->userRepository->save();
     }
 
-    public function login(string $username, string $password): void
+    public function loginUser(string $username, string $password): void
     {
         $this->validateUsernameAndPasswordFields($username, $password);
 
@@ -55,12 +59,12 @@ class AuthenticationService
             $user = $this->userRepository->getByUsername($username);
         } catch (NotFoundException $e) {
             // user does not exist
-            throw new \RuntimeException('Username does not exist');
+            throw new InvalidArgumentException('Username does not exist');
         }
 
         if (!password_verify($password, $user->getPassword())) {
             // password is incorrect
-            throw new \RuntimeException('Password is incorrect');
+            throw new InvalidArgumentException('Password is incorrect');
         }
 
         // Generate a random prefix
@@ -80,18 +84,22 @@ class AuthenticationService
 
         /** @var CacheItem $item */
         $item->set($sessionData);
-        $item->expiresAfter(86400);
+        $item->expiresAfter(self::DEFAULT_SESSION_EXPIRY_TIME);
         $this->cache->save($item);
 
         Session::put(self::SESSION_ID, $sessionId);
     }
 
-    public function logout(): void
+    public function logoutUser(): void
     {
         $sessionId = Session::get(self::SESSION_ID);
 
-        // deletes stored userId, username & privilegeLevel from cache
-        $this->cache->delete($sessionId);
+        /** @var CacheItem $cacheItem */
+        $cacheItem = $this->cache->getItem($sessionId);
+        if ($cacheItem->isHit()) {
+            // deletes stored userId, username & privilegeLevel from cache
+            $this->cache->delete($sessionId);
+        }
 
         Session::destroy();
     }
@@ -100,13 +108,13 @@ class AuthenticationService
     {
         // Ensure either of the fields are NOT empty
         if ($username === '' || $password === '') {
-            throw new \RuntimeException('Username or Password was not provided');
+            throw new InvalidArgumentException('Username or Password was not provided', 406);
         }
 
         // Make sure that the username length is at least 4 characters long
         $minUsernameLength = 4;
         if (strlen($username) < $minUsernameLength) {
-            throw new \RuntimeException("The username must be at least {$minUsernameLength} characters in length");
+            throw new LengthException("The username must be at least {$minUsernameLength} characters in length", 406);
         }
     }
 
@@ -139,7 +147,7 @@ class AuthenticationService
      * @param int $requiredPrivilegeLevel
      * @return bool
      */
-    public function hasPrivilege(int $requiredPrivilegeLevel): bool
+    public function userHasPrivilege(int $requiredPrivilegeLevel): bool
     {
         $sessionId = Session::get(self::SESSION_ID);
         if (!$sessionId) {
@@ -152,5 +160,10 @@ class AuthenticationService
         $userPrivilege = $sessionData[self::SESSION_DATA_PRIVILEGE_LEVEL] ?? 0;
 
         return ($userPrivilege === $requiredPrivilegeLevel);
+    }
+
+    public function generateRandomString(): string
+    {
+        return md5(random_bytes(10));
     }
 }
